Enhace pinchzoom
It would be better for pinch zoom to be scaled based on the length between
fingers only, rather than depending on the speed value.

A new option was added **realPinch** defaults to **false**. when true
the scale value will be calculated only from the distance between
fingers which will give an effect similar to image zoomin on
mobile phones.