Release text selection interceptor when paused
text selection interceptor is used to prevent the browser
from responsding to "dragstart" and "dragend" events during
panning.

Normally interceptor is turned on when panning is about to
start and is turned off when panning is finished.

With addition of `pause/resume` API, it is possible to get
into state when interceptro is turned on, but never turned off.

This change releases the interceptor when we unsubscribe
from events.

Should fix https://github.com/anvaka/panzoom/issues/68