Round transform coordinates to avoid blurry lines in svg
Before this patch you get the following matrix when zooming to 100%:

`matrix(0.9991895111944822 0 0 0.9991895111944822 115 -1719)`

That produces in SVG very blurry lines. After this patch the following matrix is produced:

`matrix(1 0 0 1 115 -1736)`

Which makes everything crystal  clear. A rounding range of 500 is chosen for `scale` to get smooth animations while also getting integers when zooming to 100%.